remove wandb as a main dependency
